+++ dispatch/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimize stack usage of `Dispatch::poll`
 - Replace `trussed` dependency with `trussed-core`.
 - Move the `app` and `command` modules into a separate crate, `ctaphid-app`, and re-export it.
+- Make `App` trait generic over the response size.
 
 ## [0.1.1] - 2022-08-22
 - adjust to `interchange` API change